video_ingestion/requirements.txt
pyzmq>=25.1
python-dotenv>=1.0
numpy>=1.24
psutil>=5.9
opencv-python-headless>=4.8
